# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SessionMode = (typeof SessionMode)[keyof typeof SessionMode];

export const SessionMode = {
  sitting: "sitting",
  standing: "standing",
  resting: "resting",
  walking: "walking",
} as const;

export type RestType = (typeof RestType)[keyof typeof RestType] | null;

export const RestType = {
  nap: "nap",
  sleep: "sleep",
} as const;

export interface Session {
  id: number;
  mode: SessionMode;
  startedAt: string;
  endedAt: string | null;
  durationSeconds: number | null;
  restType: RestType | null;
}

export interface SessionList {
  sessions: Session[];
  total: number;
}

export interface ActiveSessionResponse {
  session: Session | null;
}

export interface StartSessionBody {
  mode: SessionMode;
  /** Optional override for session start time (used for offline sync) */
  startedAt?: string;
}

export interface EndSessionBody {
  /** Optional override for session end time (used for offline sync) */
  endedAt?: string;
}

export interface Settings {
  id: number;
  dailyStandingGoalMinutes: number;
  sittingAlertMinutes: number;
  standingMinMinutes: number;
  standingMaxMinutes: number;
  reminderIntervalMinutes: number;
  remindersCount: number;
  autoDetectWalking: boolean;
}

export interface UpdateSettingsBody {
  dailyStandingGoalMinutes?: number;
  sittingAlertMinutes?: number;
  standingMinMinutes?: number;
  standingMaxMinutes?: number;
  reminderIntervalMinutes?: number;
  remindersCount?: number;
  autoDetectWalking?: boolean;
}

export interface TodayStats {
  date: string;
  sittingMinutes: number;
  standingMinutes: number;
  walkingMinutes: number;
  restingMinutes: number;
  activeMinutes: number;
  goalMinutes: number;
  goalProgressPercent: number;
  sessionCount: number;
  currentStreak: number;
}

export interface DayStats {
  date: string;
  sittingMinutes: number;
  standingMinutes: number;
  walkingMinutes: number;
  restingMinutes: number;
  activeMinutes: number;
  goalProgressPercent: number;
}

export interface WeeklyStats {
  days: DayStats[];
  weeklyStandingMinutes: number;
  weeklyGoalMinutes: number;
  currentStreak: number;
}

export interface DailyMetric {
  date: string;
  sittingMinutes: number;
  standingMinutes: number;
  walkingMinutes: number;
  napMinutes: number;
  sleepMinutes: number;
  activeMinutes: number;
  goalProgressPercent: number;
  healthScore: number;
}

export interface DailyMetricsResponse {
  days: DailyMetric[];
  goalMinutes: number;
}

export interface SleepConsistency {
  avgSleepStartFormatted: string;
  stddevMinutes: number;
  sampleCount: number;
}

export interface SummaryMetrics {
  currentStreak: number;
  longestStreak: number;
  weeklyAverageStandingMinutes: number;
  weeklyWalkingMinutes: number;
  weeklyAverageWalkingMinutes: number;
  bestDayDate: string | null;
  bestDayMinutes: number;
  worstDayDate: string | null;
  worstDayMinutes: number;
  healthScore: number;
  healthLabel: string;
  sleepConsistency: SleepConsistency | null;
  avgNapDurationMinutes: number;
  avgSleepDurationMinutes: number;
}

export interface FitbitStatus {
  connected: boolean;
  expiresAt?: string | null;
  connectedAt?: string | null;
}

export interface FitbitAuthUrl {
  url: string;
}

export interface FitbitStepMinute {
  /** HH:MM:SS */
  time: string;
  steps: number;
}

/**
 * Derived signal from last 3 minutes of step data
 */
export type FitbitIntradayDataSignal =
  (typeof FitbitIntradayDataSignal)[keyof typeof FitbitIntradayDataSignal];

export const FitbitIntradayDataSignal = {
  sitting: "sitting",
  standing: "standing",
  walking: "walking",
  unknown: "unknown",
} as const;

export interface FitbitIntradayData {
  minutes: FitbitStepMinute[];
  fetchedAt: string;
  /** Derived signal from last 3 minutes of step data */
  signal: FitbitIntradayDataSignal;
}

export interface FitbitAnalyticsSummary {
  nudgeCount: number;
  autoCorrectionCount: number;
  userAcceptedCount: number;
  userCancelledCount: number;
}

export type FitbitEventBodyEventType =
  (typeof FitbitEventBodyEventType)[keyof typeof FitbitEventBodyEventType];

export const FitbitEventBodyEventType = {
  nudge: "nudge",
  auto_correction: "auto_correction",
  user_accepted: "user_accepted",
  user_cancelled: "user_cancelled",
} as const;

export type FitbitEventBodyFromMode =
  (typeof FitbitEventBodyFromMode)[keyof typeof FitbitEventBodyFromMode];

export const FitbitEventBodyFromMode = {
  sitting: "sitting",
  standing: "standing",
  resting: "resting",
  walking: "walking",
} as const;

export type FitbitEventBodyToMode =
  (typeof FitbitEventBodyToMode)[keyof typeof FitbitEventBodyToMode];

export const FitbitEventBodyToMode = {
  sitting: "sitting",
  standing: "standing",
  resting: "resting",
  walking: "walking",
} as const;

export interface FitbitEventBody {
  eventType: FitbitEventBodyEventType;
  fromMode: FitbitEventBodyFromMode;
  toMode: FitbitEventBodyToMode;
  reason?: string;
}

export type ListSessionsParams = {
  /**
   * Filter sessions from this date (YYYY-MM-DD)
   */
  from?: string;
  /**
   * Filter sessions to this date (YYYY-MM-DD)
   */
  to?: string;
  limit?: number;
  offset?: number;
};

export type GetDailyMetricsParams = {
  /**
   * Start date (YYYY-MM-DD)
   */
  from: string;
  /**
   * End date (YYYY-MM-DD, inclusive)
   */
  to: string;
};

export type HandleFitbitCallbackParams = {
  code: string;
  state?: string;
};