# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fitbit/callback:
    get:
      operationId: handleFitbitCallback
      tags: [fitbit]
      summary: Fitbit OAuth callback — exchanges code for tokens
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect back to app after token exchange
...
  /fitbit/intraday:
    get:
      operationId: getFitbitIntraday
      tags: [fitbit]
      summary: Get last 15 minutes of intraday step data (cached)
      responses:
        "200":
          description: Intraday step data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FitbitIntradayData"
        "404":
          description: No Fitbit connection
...
  /fitbit/analytics/event:
    post:
      operationId: recordFitbitEvent
      tags: [fitbit]
      summary: Record a Fitbit assisted-mode event (nudge, correction, acceptance, cancellation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FitbitEventBody"
      responses:
        "201":
          description: Event recorded
components:
  schemas:
...
    FitbitStepMinute:
      type: object
      properties:
        time:
          type: string
          description: "HH:MM:SS"
        steps:
          type: integer
      required:
        - time
        - steps
    FitbitIntradayData:
      type: object
      properties:
        minutes:
          type: array
          items:
            $ref: "#/components/schemas/FitbitStepMinute"
        fetchedAt:
          type: string
          format: date-time
        signal:
          type: string
          enum: [sitting, standing, walking, unknown]
          description: Derived signal from last 3 minutes of step data
      required:
        - minutes
        - fetchedAt
        - signal
...
    FitbitEventBody:
      type: object
      properties:
        eventType:
          type: string
          enum: [nudge, auto_correction, user_accepted, user_cancelled]
        fromMode:
          type: string
          enum: [sitting, standing, resting, walking]
        toMode:
          type: string
          enum: [sitting, standing, resting, walking]
        reason:
          type: string
      required:
        - eventType
        - fromMode
        - toMode