# lib/api-zod/src/generated/types/fitbitEventBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FitbitEventBodyEventType } from "./fitbitEventBodyEventType";
import type { FitbitEventBodyFromMode } from "./fitbitEventBodyFromMode";
import type { FitbitEventBodyToMode } from "./fitbitEventBodyToMode";

export interface FitbitEventBody {
  eventType: FitbitEventBodyEventType;
  fromMode: FitbitEventBodyFromMode;
  toMode: FitbitEventBodyToMode;
  reason?: string;
}