# lib/api-zod/src/generated/types/fitbitIntradayData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FitbitIntradayDataSignal } from "./fitbitIntradayDataSignal";
import type { FitbitStepMinute } from "./fitbitStepMinute";

export interface FitbitIntradayData {
  minutes: FitbitStepMinute[];
  fetchedAt: Date;
  /** Derived signal from last 3 minutes of step data */
  signal: FitbitIntradayDataSignal;
}